Tree registry and Grammar/LSystem Refactoring
More work on integrating the concepts of having a LSystem being able to generate a tree of modules, for this I think we should instantiate some kind of registry to assure that the modules are not overlapping, instead of doing this process in a simulator. Due to the modularity, it should not be that difficult to keep track of modules in the body and assure that none of them overlap.