Refactoring Representation and Factories
Was not happy with the previous version of the Representation due to the implied mutation and recombination dependencies. Now with detaching the initialization and Representation allows for better dynamic initialization. One problem is that everything is using an initialize method, but within the evosphere it is not really that of a problem, only when unit testing.